Make counters Double instead of Int

+++ Sources/Benchmark/BenchmarkResult.swift
@@ -16,10 +16,10 @@ public struct BenchmarkResult {
     public let benchmarkName: String
     public let suiteName: String
     public let measurements: [Double]
-    public let counters: [String: Int]
+    public let counters: [String: Double]
 
     public init(
-        benchmarkName: String, suiteName: String, measurements: [Double], counters: [String: Int]
+        benchmarkName: String, suiteName: String, measurements: [Double], counters: [String: Double]
     ) {
         self.benchmarkName = benchmarkName
         self.suiteName = suiteName
